Move app::job_dot_scroll to StagesGraph::scroll_position, for encapsulation
