+++ ballista-cli/src/tui/domain/jobs.rs
@@ -88,25 +88,6 @@ pub struct JobDetails {
     pub stage_plan: Option<String>,
 }
 
-#[derive(Clone, Debug)]
-pub struct GraphNode {
-    pub id: String,
-    pub label: String,
-}
-
-#[derive(Clone, Debug)]
-pub struct GraphStage {
-    pub label: String,
-    pub nodes: Vec<GraphNode>,
-}
-
-#[derive(Clone, Debug)]
-pub struct StagesGraph {
-    pub job_id: String,
-    pub stages: Vec<GraphStage>,
-    pub edges: Vec<(String, String)>,
-}
-
 impl JobsData {
     pub fn sort_jobs(&self, jobs: &mut Vec<&Job>) {
         match self.sort_column {
