+++ ballista-cli/src/tui/mod.rs
@@ -110,7 +110,6 @@ pub async fn tui_main() -> TuiResult<()> {
                     }
                     UiData::JobStagesGraph(graph) => {
                         app.job_dot_popup = Some(graph);
-                        app.job_dot_scroll = 0;
                     }
                     UiData::JobStagesData(job_id, stages) => {
                         app.job_stages_popup = Some(JobStagesPopup::new(job_id, stages));
